requirements.dev.txt
bandit==1.7.4
black==23.1.0
coverage==6.2
django-types==0.17.0
djangorestframework-stubs==1.8.0
djlint==1.19.14
factory-boy==3.2.1
Faker==11.3.0
flake8==5.0.4
flake8-django==1.1.5
isort==5.12.0
pyright==1.1.296
pre-commit==2.15.0
responses==0.17.0
types-requests==2.28.11.13